[WP] Address pprof-label review feedback
Move the Go pprof-label structs out of structs/process.h into their own
structs/span_context.h, so the OTel span structs can join them later.

Skip 0 when the u32 snapshot counter wraps: 0 is the "invalid id"
sentinel every caller tests against, so returning it on wrap-around would
silently drop a snapshot.

Shrink go_labels_scratch_t from 616 to 304 bytes per CPU: only one
{key, value} string header pair is ever in flight, and the map bucket
(Go <1.24) and slice header (Go >=1.24) formats are mutually exclusive
per binary, so they can share storage.

Finally, check the error from the go_labels_procs map lookup, like the
five sibling lookups around it.
